Refactor: Move tests around to accommodate new performance testing
