+++ .gitignore
@@ -98,3 +98,4 @@ ENV/
 *.txt
 
 /venv--*
+/tmp-performance
